refactor: 과도한 추상화 문제 개선
